UsersComponent has been renamed ContactsComponent.
