+++ projects/mysadna/src/app/app-routing.module.ts
@@ -1,13 +1,10 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { UsersComponent } from './components/users/users.component';
-import { UserDetailsComponent } from './components/user-details/user-details.component';
-import { ProjectsComponent } from './components/projects/projects.component';
-import { ProjectDetailsComponent } from './components/project-details/project-details.component';
+import { ContactsComponent, UserDetailsComponent, ProjectsComponent, ProjectDetailsComponent } from '@/components';
 
 const routes: Routes = [
-  { path: 'users', component: UsersComponent },
+  { path: 'contacts', component: ContactsComponent },
   { path: 'user-details/:userId', component: UserDetailsComponent },
   { path: 'projects', component: ProjectsComponent },
   { path: 'project-details/:projectId', component: ProjectDetailsComponent },
