+++ projects/mysadna/src/app/app.component.html
@@ -7,7 +7,7 @@
     <div class="app-navigation-bar">
       <nav class="app-navigation">
         <a [routerLink]="'/projects'">Projects</a>
-        <a [routerLink]="'/users'">Users</a>
+        <a [routerLink]="'/contacts'">Contacts</a>
       </nav>
     </div>
   </div>
