+++ projects/mysadna/src/app/app.component.scss
@@ -1,5 +1,6 @@
 .app-container {
   width: 100%;
+  font-family: 'OpenSans', sans-serif;
 }
 
 .app-header {
@@ -15,7 +16,6 @@
     }
     >.my-sadna {
       margin: 1em;
-      font-family: OpenSans;
       font-size: 1.50em;
       font-weight: 800;
       color: #0097e1;
@@ -26,7 +26,6 @@
     >.app-navigation {
       >a {
         margin: 1em;
-        font-family: OpenSans;
         font-size: 1.25em;
         font-weight: bold;
         color: #000000;
