+++ projects/mysadna/src/app/components/contacts/contacts.component.html
@@ -1,7 +1,7 @@
-<div class="users">
-  <h2 class="users-header">Users</h2>
-  <div class="users-container">
-    <ul class="users-list">
+<div class="contacts">
+  <h2 class="contacts-header">Contacts</h2>
+  <div class="contacts-container">
+    <ul class="contacts-list">
       <li *ngFor="let user of users" class="user">
         <a [routerLink]="'/user-details/' + user.userId">
           <div class="user-image">
