+++ projects/mysadna/src/app/components/contacts/index.ts
@@ -1,0 +1,1 @@
+export * from './contacts.component';
