--- projects/mysadna/src/app/components/users/index.ts
@@ -1,1 +1,0 @@
-export * from './users.component';
